+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_training_async.py
@@ -11,6 +11,7 @@ from azure.core.exceptions import ClientAuthenticationError, HttpResponseError
 from azure.ai.formrecognizer._generated.models import Model
 from azure.ai.formrecognizer._models import CustomFormModel
 from azure.ai.formrecognizer.aio import FormTrainingClient
+from azure.ai.formrecognizer import _models
 from preparers import FormRecognizerPreparer
 from asynctestcase import AsyncFormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
@@ -130,6 +131,12 @@ class TestTrainingAsync(AsyncFormRecognizerTest):
         raw_model = raw_response[0]
         custom_model = raw_response[1]
         self.assertModelTransformCorrect(custom_model, raw_model, unlabeled=True)
+
+        custom_model_dict = custom_model.to_dict()
+        
+        custom_model_from_dict = _models.CustomFormModel.from_dict(custom_model_dict)
+        self.assertEqual(custom_model_from_dict.model_name, custom_model.model_name)
+        self.assertModelTransformCorrect(custom_model_from_dict, raw_model, unlabeled=True)
 
     @FormRecognizerPreparer()
     @GlobalClientPreparer()